+++ src/cWeb.ts
@@ -4,7 +4,8 @@ import settings from './settings';
 import webview from './webview';
 
 export default class cWeb extends webview {
-    constructor(extensionPath: string, tid: string) {
+    constructor(extensionPath: string, tid: string, homework: boolean = false) {
+        const type: string = homework ? 'homework' : 'contest';
         super({
             name: 'contest',
             extensionPath,
@@ -12,8 +13,7 @@ export default class cWeb extends webview {
             getTitle: () => `T${tid}`,
             fetchData: (postMessage, addTempFile, parseMarkdown, dispose) => {
                 new fetch({
-                    path: `/d/${settings.domain}/contest/${tid}`
-                    , addCookie: true
+                    path: `/d/${settings.domain}/${type}/${tid}`, addCookie: true
                 }).start().then(async (contestDetail) => {
                     if (contestDetail?.json !== undefined) {
                         const data = contestDetail.json;
@@ -31,8 +31,7 @@ export default class cWeb extends webview {
                     io.error(e.message);
                 });
                 new fetch({
-                    path: `/d/${settings.domain}/contest/${tid}/scoreboard`
-                    , addCookie: true
+                    path: `/d/${settings.domain}/${type}/${tid}/scoreboard`, addCookie: true
                 }).start().then(async (contestDetail) => {
                     if (contestDetail?.json !== undefined) {
                         const message = {