+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the "cyezoi" extension will be documented in this file.
 ## v0.0.25
 
 - **Enhance the navigation**
+- Fix opening homework
 - Refactor contest and problem pages to utilize dynamic tab and button registration
 - Display score on problem tree view